fix: move utility function to own file
