+++ wrapper/src/main/java/io/yubicolabs/wwwwallet/credentials/Utillities.kt
@@ -1,0 +1,8 @@
+package io.yubicolabs.wwwwallet.credentials
+
+fun getClientOptions(
+    type: String,
+    challenge: String,
+    origin: String,
+) = "{\"type\":\"$type\",\"challenge\":\"$challenge\",\"origin\":\"https://$origin\"}"
+    .toByteArray()
