+++ wrapper/src/main/java/io/yubicolabs/wwwwallet/credentials/YubicoContainer.kt
@@ -501,10 +501,3 @@ private val JSONObject.rp: JSONObject?
 
 private val JSONObject.id: String?
     get() = getOrNull("id")
-
-fun getClientOptions(
-    type: String,
-    challenge: String,
-    origin: String,
-) = "{\"type\":\"$type\",\"challenge\":\"$challenge\",\"origin\":\"https://$origin\"}"
-    .toByteArray()
